Fix servant-client tests to properly use hspec-discover.
For some reason hspec-discover wasn't enabled for servant-client,
causing test cases to be skipped and fasely reporting servant-client as
test-passing whilst it was actually supposed to fail the tests.

For example, we redefined BaseUrl recently to have a path field, but the
tests don't reflect that yet but passed anyway. Because the test case
wasn't even discovered and run.